Revert "Add medication and therapy speech data"
