+++ docs/index.html
@@ -21,8 +21,6 @@ Date: December 2024
   <script src="fallback-data.js"></script>
   <!-- Basic DSM-5 disorder summaries -->
   <script src="dsm5-data.js"></script>
-  <script src="medications-data.js"></script>
-  <script src="therapy-speech.js"></script>
   <!-- AI Chat functionality - depends on WebLLM and fallback-data being loaded first -->
   <script src="ai.js"></script>
   <!-- CSS STYLES WILL GO HERE -->
